+++ backend/internal/notes/service.go
@@ -198,7 +198,7 @@ func (s *Service) AppendToNoteContent(ctx context.Context, userID pgtype.UUID, n
 		_, err := s.repo.InsertNode(ctx, sqlcgen.InsertNodeParams{
 			ID:       node.ID,
 			NoteID:   noteID,
-			Position: int32(startPos + i),
+			Position: float64(startPos + i),
 			Type:     node.Type,
 			Data:     node.Data,
 		})
@@ -213,7 +213,7 @@ func (s *Service) AppendToNoteContent(ctx context.Context, userID pgtype.UUID, n
 				Title:      node.Text,
 				DueDate:    pgtype.Date{Valid: false},
 				Recurrence: pgtype.Text{Valid: false},
-				Position:   int32(startPos + i),
+				Position:   float64(startPos + i),
 				NodeID:     node.ID,
 			})
 			if err != nil {
@@ -561,7 +561,7 @@ func (s *Service) overwriteNoteNodes(ctx context.Context, userID pgtype.UUID, no
 		_, err := s.repo.InsertNode(ctx, sqlcgen.InsertNodeParams{
 			ID:       node.ID,
 			NoteID:   noteID,
-			Position: int32(i),
+			Position: float64(i),
 			Type:     node.Type,
 			Data:     node.Data,
 		})
@@ -576,7 +576,7 @@ func (s *Service) overwriteNoteNodes(ctx context.Context, userID pgtype.UUID, no
 				Title:      node.Text,
 				DueDate:    pgtype.Date{Valid: false},
 				Recurrence: pgtype.Text{Valid: false},
-				Position:   int32(i),
+				Position:   float64(i),
 				NodeID:     node.ID,
 			})
 			if err != nil {
